+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix mistakes in depfile for Python scripts.
+
 ## [0.7.0] - 2023-04-17
 
 ### Fixed